+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Introduce a new setting AWS_BASE_NAME to prefix all AWS settings in the project.
+- Use sentry to track errors in the front application.
 
 ### Changed